==> MVP_0/2-Busqueda_Alimentos.ipynb <==
# Imports y definiciones

import os
import json
import numpy as np
import pandas as pd
from ast import literal_eval
from collections import defaultdict

databases = 'C:/Users/Chris-Brota/Desktop/Databases'
display(os.listdir(databases))

# Se elige el JSON
db = f'{databases}/{os.listdir(databases)[1]}'

# Búsqueda de alimentos energéticos por descripción

# Definición de la palabra clave de la búsqueda
keyword = 'energ'

with open(f'{db}/Food.json', 'r', encoding='utf8') as file:
    # Definición de la lista de alimentos energéticos 
    energetic_foods = list()
    
    # Para cada línea en el archivo
    for line in file:
        data = json.loads(line.strip())
        
        if data['description'] is not None:
            if keyword in data['description']:
                energetic_foods.append((data['id'], data['name'], data['public_id']))

energetic_foods

energetic_list = ['FOOD00336', 'FOOD00336', 'FOOD00168', 'FOOD00058', 'FOOD00572', 
                  'FOOD00068', 'FOOD00367', 'FOOD00219', 'FOOD00723', 'FOOD00206', 
                  'FOOD00555', 'FOOD00109', 'FOOD00379', 'FOOD00159', 'FOOD00647', 
                  'FOOD00165', 'FOOD00548', 'FOOD00183', 'FOOD00040']

# Agregando las encontradas anteriormente (Triticale, Root vegetables, Guarana)
energetic_list.extend(['FOOD00246', 'FOOD00873', 'FOOD00915'])

# Búsqueda de alimentos energéticos por nutrientes o principios activos

# Definición de la lista de componentes
compound = ['FDB003010', 'FDB009020', 'FDB000571', 'FDB016592', 'FDB013228', 'FDB012292', 
            'FDB014504', 'FDB001108', 'FDB023903', 'FDB016251', 'FDB000572', 'FDB012738', 
            'FDB013567', 'FDB003518', 'FDB024414', 'FDB003521', 'FDB031212', 'FDB011795', 
            'FDB003523', 'FDB003191', 'FDB000455', 'FDB000453', 'FDB014601', 'FDB014797', 
            'FDB002251', 'FDB022939', 'FDB001014', 'FDB000574', 'FDB017534', 'FDB023125', 
            'FDB001977', 'FDB003729']

# Pasando a número
with open(f'{db}/Compound.json', 'r', encoding='utf8') as file:
    # Definición de la lista de las componentes
    compound_num = list()
    
    # Para cada línea en el archivo
    for line in file:
        data = json.loads(line.strip())
        
        if data['public_id'] in compound:
            compound_num.append((data['id'], data['public_id']))

# Obtener el id de la comida
comp_only_numId = [i[0] for i in compound_num]
print(comp_only_numId)

energetic_list_ppo = list()

# Revisando ahora en la base de contenidos
with open(f'{db}/Content.json', 'r', encoding='utf8') as file:
    # Definición de la lista de alimentos energéticos 
    energetic_foods_num = list()
    
    # Para cada línea en el archivo
    for line in file:
        data = json.loads(line.strip())
            
        if data['source_id'] in comp_only_numId:
            energetic_foods_num.append(data['food_id'])

len(set(energetic_foods_num))

### Resultado

No conviene buscar solo por la existencia de componentes, ya que en la base de datos se presentan incluso las que están en dosis extremadamente bajas. 

**Idea futura**: Filtrar por un rango de valores de cada componente en el alimento.

# Obtener las tablas de cada alimento
