==> MVP_0/3-Obtener_Food_Info.ipynb <==
# Imports y definiciones

import os
import json
import numpy as np
import pandas as pd
from ast import literal_eval
from collections import defaultdict

databases = 'C:/Users/Chris-Brota/Desktop/Databases'
display(os.listdir(databases))

# Se elige el JSON
db = f'{databases}/{os.listdir(databases)[1]}'
db

# Definición de los alimentos a usar

# Definición de los alimentos energéticos
food_list_publicID = ['FOOD00336', 'FOOD00168', 'FOOD00058', 'FOOD00572', 'FOOD00068', 
                      'FOOD00367', 'FOOD00219', 'FOOD00723', 'FOOD00206', 'FOOD00555', 
                      'FOOD00109', 'FOOD00379', 'FOOD00159', 'FOOD00647', 'FOOD00165', 
                      'FOOD00548', 'FOOD00183', 'FOOD00040', 'FOOD00246', 'FOOD00873', 
                      'FOOD00915']

# Copiando la lista para obtener los números
food_list_ID = [i for i in food_list_publicID]

# Pasando a número
with open(f'{db}/Food.json', 'r', encoding='utf8') as file:
    for line in file:
        # Transformando el dato a diccionario
        dict_to_rev = json.loads(line.strip())
        
        if dict_to_rev['public_id'] in food_list_ID:
            # Reemplazando en la lista
            food_list_ID = [i if dict_to_rev['public_id'] != i 
                            else dict_to_rev['id']
                            for i in food_list_ID]

# Definición del diccionario de componentes a buscar

# Definición del diccionario y lista de interés
compound_dict = dict()
compound_nums = list()
compound_list_publicID = list()

# Obteniendo la información de los componentes a utilizar
with open('interest_drink_comps_codes.csv', 'r', encoding='utf8') as file:
    for line in file:
        # Formateando la línea
        data = line.strip().split(';')
        # Agregando al diccionario
        compound_dict[data[0]] = data[1:]
        # Y a la lista
        compound_list_publicID.append(data[2])

# Definición de una lista para pasarlos a números ID
compound_list_ID = [i for i in compound_list_publicID]

# Pasando a número los "compounds"
with open(f'{db}/Compound.json', 'r', encoding='utf8') as file:
    for line in file:
        # Transformando el dato a diccionario
        dict_to_rev = json.loads(line.strip())
        
        if dict_to_rev['public_id'] in compound_list_ID:
            # Reemplazando en la lista
            compound_list_ID = [i if dict_to_rev['public_id'] != i 
                                else (dict_to_rev['id'], 'Compound')
                                for i in compound_list_ID]
            
# Pasando a número los "compounds" que no aparecen en la lista
with open(f'{db}/AccessionNumber.json', 'r', encoding='utf8') as file:
    for line in file:
        # Transformando el dato a diccionario
        dict_to_rev = json.loads(line.strip())
        
        if dict_to_rev['number'] in compound_list_ID:
            # Reemplazando en la lista
            compound_list_ID = [i if dict_to_rev['number'] != i 
                                else (dict_to_rev['compound_id'], 'Compound')
                                for i in compound_list_ID]

# Finalmente, pasando a número los "nutrients"
with open(f'{db}/Nutrient.json', 'r', encoding='utf8') as file:
    for line in file:
        # Transformando el dato a diccionario
        dict_to_rev = json.loads(line.strip())
        
        if dict_to_rev['public_id'] in compound_list_ID:
            # Reemplazando en la lista
            compound_list_ID = [i if dict_to_rev['public_id'] != i 
                                else (dict_to_rev['id'], 'Nutrient')
                                for i in compound_list_ID]

compound_dict

compound_list_ID

# Registrar las composiciones de cada alimento de interés

# Creación del diccionario con cada uno de los 
food_dict = defaultdict(list)

# Para cada alimento (food_list), revisar cada uno de los componentes
# de interés (compound_list)
for food_id in food_list_ID:
    print(f'Reviewing {food_id}...')
    
    # Definición de las entradas a revisar
    entries_desired = list()
    
    for comp_info in compound_list_ID:
        # Obteniendo los valores de las componentes
        source_id, source_type = comp_info
        
        # Generando una entrada deseada
        entries_desired.append((food_id, source_id, source_type))
    
    with open(f'{db}/Content.json', 'r', encoding='utf8') as file:
        for line in file:
            # Transformando la línea a diccionario
            dataline = json.loads(line.strip())

            # Generando una entrada a revisar para validar
            entry_to_rev = (dataline['food_id'], dataline['source_id'], 
                            dataline['source_type'])

            # Si es que esta entrada se encuentra en las de interés 
            if entry_to_rev in entries_desired:
                food_dict[food_id].append(dataline)

with open('summary/food_compounds.txt', 'w', encoding='utf8') as file:
    file.write(str(food_dict))
