==> MVP_0/4-Diseno_Tabla_Alimentos.ipynb <==
# Imports y definiciones

import os, re, json
import numpy as np
import pandas as pd
from ast import literal_eval
from collections import defaultdict

databases = 'C:/Users/Chris/Desktop/Databases'
display(os.listdir(databases))

# Se elige el JSON
db = f'{databases}/{os.listdir(databases)[0]}'
db

# Definición del diccionario de componentes a buscar (modificado en comparación con el script anterior)

# Definición del diccionario y lista de interés
compound_dict = dict()
compound_nums = list()
compound_list_publicID = list()

# Obteniendo la información de los componentes a utilizar
with open('interest_drink_comps_codes.csv', 'r', encoding='utf8') as file:
    for line in file:
        # Formateando la línea
        data = line.strip().split(';')
        # Agregando al diccionario
        compound_dict[data[0]] = data[1:]
        # Y a la lista
        compound_list_publicID.append((data[0], data[2]))

# Definición de una lista para pasarlos a números ID
compound_list_ID = [i for i in compound_list_publicID]

# Pasando a número los "compounds"
with open(f'{db}/Compound.json', 'r', encoding='utf8') as file:
    for line in file:
        # Transformando el dato a diccionario
        dict_to_rev = json.loads(line.strip())
        
        if dict_to_rev['public_id'] in list(zip(*compound_list_ID))[1]:
            # Reemplazando en la lista
            compound_list_ID = [i if dict_to_rev['public_id'] != i[1] 
                                else (i[0], dict_to_rev['id'], 'Compound')
                                for i in compound_list_ID]
            
# Pasando a número los "compounds" que no aparecen en la lista
with open(f'{db}/AccessionNumber.json', 'r', encoding='utf8') as file:
    for line in file:
        # Transformando el dato a diccionario
        dict_to_rev = json.loads(line.strip())
        
        if dict_to_rev['number'] in list(zip(*compound_list_ID))[1]:
            # Reemplazando en la lista
            compound_list_ID = [i if dict_to_rev['number'] != i[1] 
                                else (i[0], dict_to_rev['compound_id'], 'Compound')
                                for i in compound_list_ID]

# Finalmente, pasando a número los "nutrients"
with open(f'{db}/Nutrient.json', 'r', encoding='utf8') as file:
    for line in file:
        # Transformando el dato a diccionario
        dict_to_rev = json.loads(line.strip())
        
        if dict_to_rev['public_id'] in list(zip(*compound_list_ID))[1]:
            # Reemplazando en la lista
            compound_list_ID = [i if dict_to_rev['public_id'] != i[1] 
                                else (i[0], dict_to_rev['id'], 'Nutrient')
                                for i in compound_list_ID]

compound_dict

compound_list_publicID

compound_list_ID

# Creación de las tablas en bruto

with open('summary/food_compounds.txt', 'r', encoding='utf8') as file:
    food_dict = file.readline().strip()

# Modificando el string para parsearlo adecuadamente
pattern = re.compile("^defaultdict\(<class '(\w+)'>")
class_to = pattern.findall(food_dict)[0]
food_dict = food_dict.replace(f"<class '{class_to}'>", class_to)
food_dict = eval(food_dict)

# Definición del diccionario de la tabla para cada alimento
table_data = dict()

for food_id in food_dict.keys():
    # Definición del diccionario de la tabla para cada componente
    compound_data = dict()
    
    for comp_info in compound_list_ID:
        # Obtener los datos de interés
        comp_name, source_id, source_type = comp_info
        
        # Definición de una lista que almacenará los datos para 
        # la comida "food_id" con componente "source_id"
        orig_content_list_i = list()
        
        # Revisando en la información filtrada 
        for dataline in food_dict[food_id]:
            if (dataline['source_id'] == source_id and 
                dataline['source_type'] == source_type and
                dataline['orig_unit'] == 'mg/100 g'): # Filtro de unidad
                # Se obtienen los componentes de interés
                if dataline['orig_content'] is not None:
                    orig_content_list_i.append((float(dataline['orig_content']), 
                                                dataline['orig_min'], 
                                                dataline['orig_max'], 
                                                dataline['orig_unit']))
                else:
                    orig_content_list_i.append((None, 
                                                dataline['orig_min'], 
                                                dataline['orig_max'], 
                                                dataline['orig_unit']))
                    
        
        # Agregando al diccionario de componentes
        compound_data[comp_name] = orig_content_list_i
    
    # Agregando a los datos de cada alimento
    table_data[food_id] = compound_data

with open('summary/raw_food_compounds_dict.txt', 'w', encoding='utf8') as file:
    file.write(str(table_data))

# Sanidad para asegurarnos de que todos estén en mg/100g

units_total = set()

for food_id in table_data.keys():
    for comp_name in table_data[food_id]:
        for i in set([i[3] for i in table_data[food_id][comp_name]]):
            units_total.add(i)

units_total

# Creando un diccionario de traducción food_id $\rightarrow$ food_name

with open('summary/raw_food_compounds_dict.txt', 'r', encoding='utf8') as file:
    table_data = literal_eval(file.readline().strip())

# Definición del diccionario de alimentos
food_translate_dict = dict()

# Pasando a número
with open(f'{db}/Food.json', 'r', encoding='utf8') as file:
    for line in file:
        # Transformando el dato a diccionario
        dict_to_rev = json.loads(line.strip())
        
        if dict_to_rev['id'] in table_data.keys():
            # Reemplazando en la lista
            food_translate_dict[dict_to_rev['id']] = dict_to_rev['name']

food_translate_dict

# Creación de las tablas sintetizadas brutas

with open('summary/raw_food_compounds_dict.txt', 'r', encoding='utf8') as file:
    table_data = literal_eval(file.readline().strip())

# Para cada alimento
for food_id in table_data.keys():
    # Guardar la tabla
    with open(f'summary/foodTables_raw/{food_id}_{food_translate_dict[food_id]}.txt', 
              'w', encoding='utf8') as file:
        for comp_name in table_data[food_id]:
            if table_data[food_id][comp_name] != []:
                print(food_id, comp_name, np.array(table_data[food_id][comp_name])[:,0].astype(float))
                comp_value = np.mean(np.array(table_data[food_id][comp_name])[:,0].astype(float) / 1000 / 100)
                file.write(f'{comp_name};{comp_value}\n')

display(table_data[586])

# Tablas a sacar

Estas tablas en general no poseen mucha de la información relevante que se necesita, por lo cual no se considerará en el MVP0. 

dont_consider = (109, 168, 219, 905, 947)

# Filtrando las características que no aparezcan en todas las tablas

# Definición de la lista de componentes finales a considerar
compounds_mvp0 = [i[0] for i in compound_list_ID]

for filename in os.listdir('summary/foodTables_raw'):
    # Indice del alimento a revisar
    index_to_rev = int(filename.split('_')[0])
    
    # Si es que la tabla es válida, se revisan sus campos
    if not index_to_rev in dont_consider:
        with open(f'summary/foodTables_raw/{filename}', 'r', encoding='utf8') as file:
            # Definición de una lista auxiliar de los compuestos que contiene este alimento
            aux_compounds = list()
            
            for line in file:
                # Agregando los compuestos a la lista auxiliar
                aux_compounds.append(line.split(';')[0])
                
            # Finalmente se actualiza la lista solo con los componentes que se encontraron
            # en este alimento
            compounds_mvp0 = [i for i in compounds_mvp0 if i in aux_compounds]
            
print(compounds_mvp0)

# Creación de tablas finales (filtradas por alimento sin información y componentes que no aparecen en todas las tablas)

for filename in os.listdir('summary/foodTables_raw'):
    # Indice del alimento a revisar
    index_to_rev = int(filename.split('_')[0])
    
    # Si es que el alimento es válido, se revisan sus campos
    if not index_to_rev in dont_consider:
        with open(f'summary/foodTables_raw/{filename}', 'r', encoding='utf8') as file:
            # Definición de una lista auxiliar de la información a agregar en la tabla definitiva
            info_to_append = list()
            
            for line in file:
                # Agregando los compuestos a la lista auxiliar
                compound_to_rev = line.split(';')[0]
                
                # Si es que el compuesto es válido, se agrega a la lista de interés
                if compound_to_rev in compounds_mvp0:
                    info_to_append.append(line)
            
        # Finalmente, con toda la información recolectada, se escribe la tabla definitiva
        with open(f'summary/foodTables/{filename}', 'w', encoding='utf8') as file:
            file.writelines(info_to_append)
